Moving completion checks to xnat_pbs_jobs repo

+++ CheckPackageExistence/CheckPackageExistence.sh
@@ -29,6 +29,7 @@ didAnyTestsFail="False"
 
 PREPROCESSING_PACKAGE_DIR="preproc"
 PREPROC_PACKAGE_TYPES="Structural rfMRI_REST1 rfMRI_REST2 tfMRI_EMOTION tfMRI_GAMBLING tfMRI_LANGUAGE tfMRI_MOTOR tfMRI_RELATIONAL tfMRI_SOCIAL tfMRI_WM Diffusion"
+PREPROC_EXTENDED_PACKAGE_TYPES="Structural_preproc_extended"
 
 FIX_PACKAGE_DIR="fix"
 FIX_PACKAGE_TYPES="rfMRI_REST"
@@ -79,6 +80,10 @@ usage() {
     echo "      Preprocessing packages will be checked for at: "
     echo "      <packaging-root>/<subject-id>/${PREPROCESSING_PACAKGE_DIR}"
     echo ""
+	echo "    [--suppress-checksum-regen]"
+	echo "    : if specified, then the checksum regeneration check will"
+	echo "      not be done"
+	echo ""
     echo " Note:"
     echo ""
     echo "   The following preproc package related files are checked: "
@@ -87,6 +92,13 @@ usage() {
         echo "    <subject-id>_3T_${aPackageType}_preproc.zip" 
         echo "    <subject-id>_3T_${aPackageType}_preproc.zip.md5" 
     done
+
+	echo "   The following preproc extended package related files are checked: "
+	for aPackageType in ${PREPROC_EXTENDED_PACKAGE_TYPES} ; do 
+		echo "    <subject-id>_3T_${aPackageType}.zip"
+		echo "    <subject-id>_3T_${aPackageType}.zip.md5"
+	done
+
     echo ""
     echo "   The following fix package related files are checked: "
     for aPackageType in ${FIX_PACKAGE_TYPES} ; do
@@ -127,6 +139,7 @@ get_options() {
     unset debug
     unset subjectId
     rootDir=${DEFAULT_PACKAGING_ROOT}
+	g_suppress_checksum_regen="FALSE"
 
     # parse arguments
     local index=0
@@ -161,6 +174,10 @@ get_options() {
                 rootDir=${argument/*=/""}
                 index=$(( index + 1 ))
                 ;;
+            --suppress-checksum-regen)
+                g_suppress_checksum_regen="TRUE"
+                index=$(( index + 1 ))
+                ;;
             *)
                 echo "Unrecognized Option: ${argument}"
                 usage
@@ -181,6 +198,24 @@ get_options() {
     # report options
     debugEcho "${FUNCNAME}: subjectId: ${subjectId}"
     debugEcho "${FUNCNAME}: rootDir: ${rootDir}"
+	debugEcho "${FUNCNAME}: g_suppress_checksum_regen: ${g_suppress_checksum_regen}"
+}
+
+get_date()
+{
+	local path=${1}
+	local __functionResultVar=${2}
+	local file_info
+	local the_date
+
+	if [ -e "${path}" ] ; then
+		file_info=`ls -lh ${path}`
+		the_date=`echo ${file_info} | cut -d " " -f 6-8`
+	else
+		the_date="DOES NOT EXIST"
+	fi
+
+	eval $__functionResultVar="'${the_date}'"
 }
 
 #
@@ -193,15 +228,22 @@ check_package_file() {
 
     local packageCheckSumFileName="${packageFileName}.md5"
     local packageFileExists="False"
+	local packageFileSize="N/A"
     local checksumFileExists="False"
     local tmpFile=""
     local checkSumsEquivalent="False"
+	local subject=""
 
     # check for package file existence
     testDesc="${packageFileName##*/}"
 
+	subject="${packageFileName%%_3T*}"
+	subject="${subject##*/}"
+
     if [ -e ${packageFileName} ]; then
         packageFileExists="True"
+		packageFileInfo=`ls -lh ${packageFileName}`
+		packageFileSize=`echo ${packageFileInfo} | cut -d " " -f 5`
     fi
 
     # check for checksum file existence
@@ -211,20 +253,27 @@ check_package_file() {
 
     # check the checksum value
     if [ "${packageFileExists}" = "True" ] && [ "${checksumFileExists}" = "True" ]; then
-        tmpFile=`mktemp`
-        md5sum ${packageFileName} | cut -d' ' -f1 > ${tmpFile}.1
-        cat ${packageCheckSumFileName} | cut -d' ' -f1 > ${tmpFile}.2
+		
+		if [ "${g_suppress_checksum_regen}" = "TRUE" ] ; then
+			checkSumsEquivalent="Not Checked"
+		else
+			tmpFile=`mktemp`
+			md5sum ${packageFileName} | cut -d' ' -f1 > ${tmpFile}.1
+			cat ${packageCheckSumFileName} | cut -d' ' -f1 > ${tmpFile}.2
             
-        if diff ${tmpFile}.1 ${tmpFile}.2 > /dev/null; then
-            checkSumsEquivalent="True"
-        fi
+			if diff ${tmpFile}.1 ${tmpFile}.2 > /dev/null; then
+				checkSumsEquivalent="True"
+			fi
+			
+			rm -f ${tmpFile}.1
+			rm -f ${tmpFile}.2
+		fi
+	fi
 
-        rm -f ${tmpFile}.1
-        rm -f ${tmpFile}.2
-    fi
+	get_date ${packageFileName} package_date
 
     # output information
-    echo -e "${testDesc}:\t${packageFileExists}\t${checksumFileExists}\t${checkSumsEquivalent}"
+    echo -e "${subject}\t${testDesc}\t${packageFileExists}\t${packageFileSize}\t${checksumFileExists}\t${checkSumsEquivalent}\t${package_date}"
 
     # return success or failure indication
     if [ "${packageFileExists}" != "True" ] || [ "${checksumFileExists}" != "True" ] || [ "${checkSumsEquivalent}" != "True" ]; then
@@ -258,7 +307,8 @@ main() {
     subjectFixExtendedPackageDir="${rootDir}/${subjectId}/${FIX_EXTENDED_PACKAGE_DIR}"
     # echo "Subject FIX Extended Package Directory: ${subjectFixPackageDir}"
 
-    echo -e "package file:\tExists\tChecksum Exists\tChecksum correct"
+	# Header line
+    #echo -e "subject\tpackage file\tExists\tSize\tChecksum Exists\tChecksum correct\tPackage Date"
 
     # check preproc packages
     for packageType in ${PREPROC_PACKAGE_TYPES}; do
@@ -268,36 +318,44 @@ main() {
         fi
     done
 
-    # check fix packages
-    for packageType in ${FIX_PACKAGE_TYPES}; do
-        packageFileName="${subjectFixPackageDir}/${subjectId}_3T_${packageType}_fix.zip"
-        if ! check_package_file ${packageFileName} ; then
-            didAnyTestsFail="True"
-        fi
-    done
+	# check preproc extended packages
+	for packageType in ${PREPROC_EXTENDED_PACKAGE_TYPES} ; do
+		packageFileName="${subjectPackageDir}/${subjectId}_3T_${packageType}.zip"
+		if ! check_package_file ${packageFileName} ; then
+			didAnyTestsFail="True"
+		fi
+	done
 
-    # check fixextended packages
-    for packageType in ${FIX_EXTENDED_PACKAGE_TYPES} ; do
-        packageFileName="${subjectFixExtendedPackageDir}/${subjectId}_3T_${packageType}_fixextended.zip"
-        if ! check_package_file ${packageFileName} ; then
-            didAnyTestsFail="True"
-        fi
-    done
+    # # check fix packages
+    # for packageType in ${FIX_PACKAGE_TYPES}; do
+    #     packageFileName="${subjectFixPackageDir}/${subjectId}_3T_${packageType}_fix.zip"
+    #     if ! check_package_file ${packageFileName} ; then
+    #         didAnyTestsFail="True"
+    #     fi
+    # done
 
-    # check task analysis packages
-    for smoothing_level in ${TASK_ANALYSIS_SMOOTHING_LEVELS} ; do
+    # # check fixextended packages
+    # for packageType in ${FIX_EXTENDED_PACKAGE_TYPES} ; do
+    #     packageFileName="${subjectFixExtendedPackageDir}/${subjectId}_3T_${packageType}_fixextended.zip"
+    #     if ! check_package_file ${packageFileName} ; then
+    #         didAnyTestsFail="True"
+    #     fi
+    # done
 
-        subjectTaskAnalysisDir="${rootDir}/${subjectId}/analysis_s${smoothing_level}"
-        # echo "Subject Task Analysis Package Directory: ${subjectTaskAnalysisDir}"
-        # echo -e "package file:\tExists\tChecksum Exists\tChecksum correct"
-        for packageType in ${TASK_ANALYSIS_PACKAGE_TYPES} ; do
-            packageFileName="${subjectTaskAnalysisDir}/${subjectId}_3T_${packageType}_analysis_s${smoothing_level}.zip"
-            if ! check_package_file ${packageFileName} ; then
-                didAnyTestsFail="True"
-            fi
-        done
+    # # check task analysis packages
+    # for smoothing_level in ${TASK_ANALYSIS_SMOOTHING_LEVELS} ; do
 
-    done
+    #     subjectTaskAnalysisDir="${rootDir}/${subjectId}/analysis_s${smoothing_level}"
+    #     # echo "Subject Task Analysis Package Directory: ${subjectTaskAnalysisDir}"
+    #     # echo -e "package file:\tExists\tChecksum Exists\tChecksum correct"
+    #     for packageType in ${TASK_ANALYSIS_PACKAGE_TYPES} ; do
+    #         packageFileName="${subjectTaskAnalysisDir}/${subjectId}_3T_${packageType}_analysis_s${smoothing_level}.zip"
+    #         if ! check_package_file ${packageFileName} ; then
+    #             didAnyTestsFail="True"
+    #         fi
+    #     done
+
+    # done
 }
 
 #
